+++ packages/unlock-js/src/walletService.ts
@@ -266,6 +266,10 @@ export default class WalletService extends UnlockService {
       data?: string
       decimals?: number
       erc20Address?: string
+      keyPrice?: string
+      recurringPayment?: string | number
+      totalApproval?: string
+      extendApprovalOnly?: boolean
     },
     transactionOptions?: TransactionOptions,
     callback?: WalletServiceCallback