+++ unlock-app/sentry.client.config.js
@@ -19,5 +19,7 @@ Sentry.init({
   debug: process.env.NODE_ENV === 'development',
   attachStacktrace: true,
   allowUrls: true,
+  enabled: ['production', 'staging'].includes(process.env.NODE_ENV),
+  environment: process.env.NODE_ENV,
   autoSessionTracking: true,
 })