+++ unlock-app/src/components/interface/keychain/Key.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useContext, Fragment } from 'react'
+import React, { useState, useContext, Fragment, MouseEventHandler } from 'react'
 import useClipboard from 'react-use-clipboard'
 import {
   AvatarImage,
@@ -17,19 +17,17 @@ import {
 } from 'react-icons/ri'
 import { Badge, Button, minifyAddress } from '@unlock-protocol/ui'
 import { networks } from '@unlock-protocol/networks'
-import { expirationAsDate } from '../../../utils/durations'
 import QRModal from './QRModal'
 import useMetadata from '../../../hooks/useMetadata'
 import { useWalletService } from '../../../utils/withWalletService'
 import WedlockServiceContext from '../../../contexts/WedlocksContext'
 import { useAuth } from '../../../contexts/AuthenticationContext'
-import { MAX_UINT } from '../../../constants'
 import { useConfig } from '../../../utils/withConfig'
 import { OpenSeaIcon } from '../../icons'
 import { CancelAndRefundModal } from './CancelAndRefundModal'
-import { KeyMetadataDrawer } from './KeyMetadataDrawer'
+import { KeyInfoDrawer } from './KeyInfoDrawer'
 import { lockTickerSymbol } from '~/utils/checkoutLockUtils'
-import { isError, useQuery } from '@tanstack/react-query'
+import { useQuery } from '@tanstack/react-query'
 import { useWeb3Service } from '~/utils/withWeb3Service'
 import { Menu, Transition } from '@headlessui/react'
 import { classed as tw } from '@tw-classed/react'
@@ -43,11 +41,9 @@ import {
   RiFileCopyLine as CopyLineIcon,
   RiExternalLinkFill as ExternalIcon,
 } from 'react-icons/ri'
-import { useStorageService } from '~/utils/withStorageService'
 import dayjs from 'dayjs'
 import { ExtendMembershipModal } from './Extend'
-import { MdOutlineAdd as IncreaseIcon } from 'react-icons/md'
-import { ethers } from 'ethers'
+import { MAX_UINT } from '~/constants'
 
 export const MenuButton = tw.button(
   'group flex gap-2 w-full font-semibold items-center rounded-md px-2 py-2 text-sm disabled:opacity-50 disabled:cursor-not-allowed',
@@ -59,137 +55,6 @@ export const MenuButton = tw.button(
     },
   }
 )
-
-interface KeyBoxProps {
-  lock: any
-  expiration: string
-  tokenId: string
-  network: number
-  expirationStatus: string
-}
-
-const KeyBoxItem = ({ label, value }: Record<'label' | 'value', string>) => {
-  return (
-    <div className="flex items-center justify-between gap-2 py-1 text-sm">
-      <span className="text-gray-500">{label}</span>
-      <span className="font-bold">{value}</span>
-    </div>
-  )
-}
-
-const KeyBox = ({
-  lock,
-  expiration,
-  tokenId,
-  network,
-  expirationStatus,
-}: KeyBoxProps) => {
-  const metadata = useMetadata(lock.address, tokenId, network)
-  const config = useConfig()
-  const [_, setCopied] = useClipboard(lock.address, {
-    successDuration: 2000,
-  })
-  const storage = useStorageService()
-  const walletService = useWalletService()
-  const { account } = useAuth()
-  const { data: subscriptionInfo } = useQuery(
-    ['subscriptions', lock.address, tokenId, network],
-    async () => {
-      storage.loginPrompt({
-        walletService,
-        address: account!,
-        chainId: 1,
-      })
-      const response = await storage.getSubscription({
-        network: network,
-        lockAddress: lock.address,
-        keyId: tokenId,
-      })
-      return response.subscriptions?.[0] ?? {}
-    },
-    {
-      retry: 2,
-      enabled: !!(lock?.address && account),
-      onError(error) {
-        console.error(error)
-      },
-    }
-  )
-
-  return (
-    <div className="grid gap-2">
-      <div>
-        <Avatar className="flex items-center justify-center ">
-          <AvatarImage
-            className="w-full h-full rounded-xl aspect-1 max-h-72 max-w-72 "
-            alt={lock.name}
-            src={metadata.image}
-            width={250}
-            height={250}
-          />
-          <AvatarFallback className="uppercase" delayMs={100}>
-            {lock.name.slice(0, 2)}
-          </AvatarFallback>
-        </Avatar>
-      </div>
-      <div className="flex items-center justify-between">
-        <div className="inline-flex items-center gap-2">
-          {minifyAddress(lock.address)}
-          <button
-            onClick={(event) => {
-              event.preventDefault()
-              setCopied()
-              ToastHelper.success('Copied!')
-            }}
-          >
-            <CopyLineIcon size={18} />
-          </button>
-        </div>
-        <a
-          href={config.networks?.[network]?.explorer?.urls.address(
-            lock.address
-          )}
-          target="_blank"
-          rel="noreferrer"
-          className="inline-flex items-center gap-2 text-ui-main-500"
-        >
-          View <ExternalIcon size={18} />
-        </a>
-      </div>
-      <div className="divide-y divide-y-reverse divide-brand-dark">
-        <h3 className="text-lg font-bold line-clamp-1">{lock.name}</h3>
-        <KeyBoxItem label="Token ID" value={tokenId} />
-        {expiration !== MAX_UINT && (
-          <KeyBoxItem label="Valid" value={expirationStatus} />
-        )}
-        <KeyBoxItem
-          label="Renews On"
-          value={
-            subscriptionInfo?.next
-              ? dayjs.unix(subscriptionInfo.next).format('D MMM YYYY, h:mm A')
-              : '-'
-          }
-        />
-        <KeyBoxItem
-          label="Renew Cycle"
-          value={subscriptionInfo?.approvedTime || '-'}
-        />
-        <KeyBoxItem
-          label="Payment Type"
-          value={subscriptionInfo?.type || '-'}
-        />
-        <KeyBoxItem
-          label="User Balance"
-          value={
-            subscriptionInfo?.balance
-              ? `${subscriptionInfo.balance.amount} ${subscriptionInfo.balance.symbol}`
-              : '-'
-          }
-        />
-      </div>
-    </div>
-  )
-}
 
 export interface KeyProps {
   id: string
@@ -212,6 +77,7 @@ export interface KeyProps {
     createdAtBlock?: any
   }
 }
+
 export interface Props {
   ownedKey: KeyProps
   account: string
@@ -226,19 +92,16 @@ function Key({ ownedKey, account, network }: Props) {
   const web3Service = useWeb3Service()
   const { watchAsset } = useAuth()
   const config = useConfig()
-  const expirationStatus = expirationAsDate(expiration)
-
-  const [error, setError] = useState<string | null>()
   const [showingQR, setShowingQR] = useState(false)
-  const [showMetadata, setShowMetadata] = useState(false)
+  const [showMoreInfo, setShowMoreInfo] = useState(false)
   const [signature, setSignature] = useState<any | null>(null)
   const [showCancelModal, setShowCancelModal] = useState(false)
   const [expireAndRefunded, setExpireAndRefunded] = useState(false)
   const [showExtendMembershipModal, setShowExtendMembership] = useState(false)
   const isKeyExpired =
-    ownedKey.expiration !== MAX_UINT
+    (ownedKey.expiration !== MAX_UINT
       ? dayjs.unix(parseInt(ownedKey.expiration)).isBefore(dayjs())
-      : false
+      : false) || expireAndRefunded
 
   const { data: lockData, isLoading: isLockDataLoading } = useQuery(
     ['lock', lock.address, network],
@@ -246,9 +109,16 @@ function Key({ ownedKey, account, network }: Props) {
       return web3Service.getLock(lock.address, network)
     }
   )
+  const metadata = useMetadata(lock.address, tokenId, network)
+  const [_, setCopied] = useClipboard(lock.address, {
+    successDuration: 2000,
+  })
 
-  const handleSignature = async () => {
-    setError('')
+  const handleSignature: MouseEventHandler<HTMLButtonElement> = async (
+    event
+  ) => {
+    event.preventDefault()
+    event.stopPropagation()
     const payload = JSON.stringify({
       network,
       account,
@@ -256,9 +126,7 @@ function Key({ ownedKey, account, network }: Props) {
       timestamp: Date.now(),
       tokenId,
     })
-
     const signature = await walletService.signMessage(payload, 'personal_sign')
-
     setSignature({
       payload,
       signature,
@@ -303,7 +171,7 @@ function Key({ ownedKey, account, network }: Props) {
         qrImage
       )
     } catch {
-      setError('We could not send the email. Please try again later')
+      ToastHelper.error('We could not send the email. Please try again later')
     }
   }
 
@@ -325,25 +193,22 @@ function Key({ ownedKey, account, network }: Props) {
     ownedKey.lock.tokenAddress !== '0x0000000000000000000000000000000000000000'
 
   const isExtendable =
-    ownedKey.lock.version >= 11 &&
-    ownedKey.expiration !== MAX_UINT &&
-    isKeyExpired
+    ownedKey.lock.version >= 11 && ownedKey.expiration !== MAX_UINT
 
   const isRenewable =
-    ownedKey.lock.version >= 11 &&
-    ownedKey.expiration !== MAX_UINT &&
-    isERC20 &&
-    !isKeyExpired
+    ownedKey.lock.version >= 11 && ownedKey.expiration !== MAX_UINT && isERC20
 
   return (
     <div className="grid gap-6 p-4 bg-white border border-gray-200 shadow-lg rounded-xl">
-      <KeyMetadataDrawer
-        isOpen={showMetadata}
-        setIsOpen={setShowMetadata}
+      <KeyInfoDrawer
+        isOpen={showMoreInfo}
+        setIsOpen={setShowMoreInfo}
         account={account}
         lock={lock}
         tokenId={tokenId}
         network={network}
+        expiration={expiration}
+        imageURL={metadata.image}
       />
       <CancelAndRefundModal
         isOpen={showCancelModal}
@@ -466,11 +331,11 @@ function Key({ ownedKey, account, network }: Props) {
                         active={active}
                         onClick={(event) => {
                           event.preventDefault()
-                          setShowMetadata(true)
+                          setShowMoreInfo(true)
                         }}
                       >
                         <InfoIcon />
-                        View metadata
+                        Show details
                       </MenuButton>
                     )}
                   </Menu.Item>
@@ -518,14 +383,55 @@ function Key({ ownedKey, account, network }: Props) {
           </Menu>
         </div>
       </div>
-      <KeyBox
-        network={network}
-        lock={lock}
-        expiration={expiration}
-        tokenId={tokenId}
-        expirationStatus={expirationStatus}
-      />
-      {error && <span className="text-sm text-red-500">{error}</span>}
+      <div className="grid gap-2">
+        <Avatar
+          onClick={(event) => {
+            event.preventDefault()
+            setShowMoreInfo(true)
+          }}
+          className="flex items-center justify-center cursor-pointer hover:bg-gray-50"
+        >
+          <AvatarImage
+            className="w-full h-full rounded-xl aspect-1 max-h-72 max-w-72"
+            alt={lock.name}
+            src={metadata.image}
+            width={250}
+            height={250}
+          />
+          <AvatarFallback
+            className="flex flex-col items-center justify-center text-3xl font-bold uppercase rounded-xl aspect-1 h-72 w-72"
+            delayMs={100}
+          >
+            {lock?.name?.slice(0, 2)}
+          </AvatarFallback>
+        </Avatar>
+        <div className="flex items-center justify-between rounded">
+          <div className="inline-flex items-center gap-2">
+            {minifyAddress(lock.address)}
+            <button
+              aria-label="Copy Lock Address"
+              onClick={(event) => {
+                event.preventDefault()
+                setCopied()
+                ToastHelper.success('Copied!')
+              }}
+            >
+              <CopyLineIcon size={18} />
+            </button>
+          </div>
+          <a
+            href={config.networks?.[network]?.explorer?.urls.address(
+              lock.address
+            )}
+            target="_blank"
+            rel="noreferrer"
+            className="inline-flex items-center gap-2 text-ui-main-500 px-2 py-0.5 rounded-lg bg-ui-main-50 hover:bg-ui-main-100 hover:text-ui-main-600"
+          >
+            View <ExternalIcon size={18} />
+          </a>
+        </div>
+        <h3 className="text-xl font-bold rounded">{lock.name}</h3>
+      </div>
     </div>
   )
 }