+++ unlock-app/src/components/interface/locks/Manage/elements/Members.tsx
@@ -3,7 +3,6 @@ import { useAuth } from '~/contexts/AuthenticationContext'
 import { useStorageService } from '~/utils/withStorageService'
 import { useWalletService } from '~/utils/withWalletService'
 import { ToastHelper } from '~/components/helpers/toast.helper'
-import { useWeb3Service } from '~/utils/withWeb3Service'
 import { ImageBar } from './ImageBar'
 import { MemberCard } from './MemberCard'
 import { paginate } from '~/utils/pagination'
@@ -11,6 +10,7 @@ import { PaginationBar } from './PaginationBar'
 import React from 'react'
 import { ExpirationStatus } from './FilterBar'
 import Link from 'next/link'
+import { subgraph } from '~/config/subgraph'
 
 interface MembersProps {
   lockAddress: string
@@ -59,7 +59,6 @@ export const Members = ({
 }: MembersProps) => {
   const { account } = useAuth()
   const walletService = useWalletService()
-  const web3Service = useWeb3Service()
   const storageService = useStorageService()
 
   const getMembers = async () => {
@@ -75,14 +74,9 @@ export const Members = ({
     })
   }
 
-  const getLockVersion = async (): Promise<number> => {
-    if (!network) return 0
-    return web3Service.publicLockVersion(lockAddress, network)
-  }
-
   const [
     { isLoading, data: members = [] },
-    { isLoading: isLoadingVersion, data: lockVersion = 0 },
+    { isLoading: isLockLoading, data: lock },
   ] = useQueries({
     queries: [
       {
@@ -92,17 +86,27 @@ export const Members = ({
           ToastHelper.error(`Can't load members, please try again`)
         },
       },
+
       {
-        queryFn: getLockVersion,
-        queryKey: ['getLockVersion', lockAddress, network],
+        queryFn: () => {
+          return subgraph.lock(
+            {
+              where: {
+                address: lockAddress,
+              },
+            },
+            { network }
+          )
+        },
+        queryKey: ['getSubgraphLock', lockAddress, network],
         onError: () => {
-          ToastHelper.error('Cant get lock version, please try again')
+          ToastHelper.error('Unable to fetch lock from subgraph')
         },
       },
     ],
   })
 
-  const loading = isLoadingVersion || isLoading || loadingFilters
+  const loading = isLockLoading || isLoading || loadingFilters
   const noItems = members?.length === 0 && !loading
 
   const hasActiveFilter =
@@ -172,10 +176,11 @@ export const Members = ({
             token={token}
             owner={owner}
             expiration={expiration}
-            version={lockVersion}
+            version={lock?.version}
             metadata={metadata}
             lockAddress={lockAddress!}
             network={network}
+            expirationDuration={lock?.expirationDuration}
           />
         )
       })}