+++ packages/unlock-js/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changes
 
+# 0.32.15
+
+- Add ability to pass recurring payment and total approval to extend key
+
+# 0.32.13
+
+- Export utility types for subgraph service
+
 # 0.32.11
 
 - add `updateTransferFee` in walletService